requirements.txt
Flask==2.2.2
requests==2.28.1
PyYAML==6.0
black==22.12.0
pandas==1.5.2